# WorkplaceViolencePredictionAPI/docs/openapi.json
openapi: 3.0.3
info:
  title: Workplace Violence Prediction API
  version: 1.0.0
  description: An automated system to predict potential incidents of workplace violence
    in a hospital
paths:
...
  /api/email/:
    get:
      operationId: email_list
      tags:
      - email
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmailRecipient'
          description: ''
    post:
      operationId: email_create
      tags:
      - email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRecipient'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EmailRecipient'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EmailRecipient'
        required: true
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailRecipient'
          description: ''
...
components:
  schemas:
...
    EmailRecipient:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          nullable: true
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 254
      required:
      - email
      - id
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: http
      scheme: bearer